Aggregate file.
